% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Anscombe.R
\docType{data}
\name{Anscombe}
\alias{Anscombe}
\title{Anscombe's Four Datasets}
\format{A data frame with 11 rows and 8 variables:
\describe{
  \item{x1}{values of \emph{X} for the first dataset}
  \item{y1}{values of \emph{Y} for the first dataset}
  \item{x2}{values of \emph{X} for the second dataset}
  \item{y2}{values of \emph{Y} for the second dataset}
  \item{x3}{values of \emph{X} for the third dataset}
  \item{y3}{values of \emph{Y} for the third dataset}
  \item{x4}{values of \emph{X} for the fourth dataset}
  \item{y4}{values of \emph{Y} for the fourth dataset}
}}
\usage{
Anscombe
}
\description{
This dataset is used to illustrate the importance of statistical display as an adjunct to summary statistics. Anscombe (1973) fabricated four different bivariate datasets such that, for all datasets, the respective \emph{X} and \emph{Y} means, \emph{X} and \emph{Y} standard deviations, and correlations, slopes, intercepts, and standard errors of estimate are equal. Accordingly, without a visual representation of these four panels, one might assume that the data values for all four datasets are the same. Scatterplots illustrate, however, the extent to which these datasets are different from one another.
}
\keyword{datasets}
