% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Basketball.R
\docType{data}
\name{Basketball}
\alias{Basketball}
\title{Heights and Weights of U.S. Basketball Players}
\format{A data frame with 20 rows and 6 variables:
\describe{
  \item{player}{name of player}
  \item{gender}{gender of player}
  \item{heightin}{height of player in inches}
  \item{weightlb}{weight of player in pounds}
  \item{games}{number of games played}
  \item{points}{average total points scored per game}
}}
\source{
\url{http://www.espn.com/}
}
\usage{
Basketball
}
\description{
The dataset consists of the heights and weights of the 24 scoring leaders, 12 each from the U.S. Women’s and Men’s National Basketball Association, for the 2014 – 2015 season. These data are taken from the ESPN website at espn.com.
}
\keyword{datasets}
