% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Blood.R
\docType{data}
\name{Blood}
\alias{Blood}
\title{Blood Pressure Data of African-American Adult Males}
\format{A data frame with 21 rows and 4 variables:
\describe{
  \item{id}{case number}
  \item{treatmen}{treatment condition}
  \item{systolc1}{initial blood pressure}
  \item{systolc2}{final blood pressure}
}}
\usage{
Blood
}
\description{
The data were collected to determine whether an increase in calcium intake reduces blood pressure among African-American adult males. The data are based on a sample of 21 African-American adult males selected randomly from the population of African-American adult males. Ten of the 21 men were randomly assigned to a treatment condition that required them to take a calcium supplement for 12 weeks. The remaining 11 men received a placebo for the 12 weeks. At both the beginning and the end of this time period, systolic blood pressure readings of all men were recorded. These data are adapted from the Data and Story Library (DASL) website.
}
\keyword{datasets}
