% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figures.R
\docType{data}
\name{Chapter14_Figures}
\alias{Chapter14_Figures}
\title{Exercise 14.1 Figures}
\format{A data frame with 12 rows and 7 variables:
\describe{
  \item{sex}{individual's sex}
  \item{score1}{reading achievement score for first scenario}
  \item{method}{teaching method}
  \item{score2}{reading achievement score for second scenario}
  \item{score3}{reading achievement score for third scenario}
  \item{score4}{reading achievement score for fourth scenario}
  \item{score5}{reading achievement score for fifth scenario}
}}
\usage{
Chapter14_Figures
}
\description{
This dataset contains simulated data for the figures accompanying Exercise 14.1 of Chapter 14. The data represent the results of a fictional study to determine whether there is a relationship between gender, teaching method, and achievement in reading. Each set of scores reflects a scenario with a different relationship among the variables.
}
\keyword{datasets}
