% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Currency.R
\docType{data}
\name{Currency}
\alias{Currency}
\title{Value and Circulation of Currency}
\format{A data frame with 5 rows and 3 variables:
\describe{
  \item{BillValue}{denomination}
  \item{TotalCirculation}{total currency in circulation in U.S. dollars}
  \item{NumberCirculation}{total number of bills in circulation}
}}
\usage{
Currency
}
\description{
This dataset contains, for the smaller bill denominations, the value of the bill and the total value in circulation. The source for these data is \emph{The World Almanac and Book of Facts 2014}.
}
\keyword{datasets}
