% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Exercise.R
\docType{data}
\name{Exercise}
\alias{Exercise}
\title{Exercise, Food Intake, and Weight Loss}
\format{A data frame with 10 rows and 4 variables:
\describe{
  \item{ID}{case number}
  \item{Exercise}{average daily number of hours exercised in that week}
  \item{FoodIntake}{average daily number of calories consumed in one particular week that is more than a baseline of 1,000 calories, as measured in increments of 100 calories}
  \item{WeightLoss}{number of pounds lost in that week}
}}
\usage{
Exercise
}
\description{
A fabricated dataset constructed by Darlington (1990) to demonstrate the importance of including all relevant variables in an analysis.  This dataset contains information about exercise, food intake, and weight loss for a fictional set of dieters.
}
\references{
"Regression and linear models." Darlington, R. B. (1990, ISBN:978-0070153721)
}
\keyword{datasets}
