% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figures.R
\docType{data}
\name{Exercise14_5}
\alias{Exercise14_5}
\title{Exercise 14.5 Data}
\format{A data frame with 40 rows and 3 variables:
\describe{
  \item{Time}{time of day student takes the course}
  \item{Year}{year of college in which the student is enrolled}
  \item{Score}{final exam score}
}}
\usage{
Exercise14_5
}
\description{
This dataset contains simulated data for the figures accompanying Exercise 14.1 of Chapter 14. The data represent the results of a fictional study in which a college professor examines the effect of the grade level of the students and the time of the course on how well undergraduate students at her college do in her course.
}
\keyword{datasets}
