% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figures.R
\docType{data}
\name{Figure15_1}
\alias{Figure15_1}
\title{Figure 15.1 Data}
\format{A list with 3 elements:
\describe{
  \item{x}{an integer-scaled independent variable}
  \item{y}{an integer-scaled outcome variable}
  \item{f}{frequency of value pair}
}}
\usage{
Figure15_1
}
\description{
This dataset contains simulated data for Figure 15.1 of Chapter 15.
}
\keyword{datasets}
