% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figures.R
\docType{data}
\name{Figure15_12}
\alias{Figure15_12}
\title{Figure 15.12 Data}
\format{A data frame with 9 rows and 4 variables:
\describe{
  \item{x}{a numeric independent variable for Figure 15.12}
  \item{y}{a numeric outcome variable for Figure 15.12}
  \item{xpr}{a numeric independent variable for Figure 15.13}
  \item{ypr}{a numeric outcome variable for Figure 15.13}
}}
\usage{
Figure15_12
}
\description{
This dataset contains simulated data for Figures 15.12 - 15.13 of Chapter 15.
}
\keyword{datasets}
