% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figures.R
\docType{data}
\name{Figure15_9}
\alias{Figure15_9}
\title{Figure 15.9 Data}
\format{A data frame with 24 rows and 4 variables:
\describe{
  \item{x}{a numeric independent variable for Figure 15.9}
  \item{y}{a numeric outcome variable for Figure 15.9}
  \item{res}{residual value for regression of \code{y} on \code{x}}
  \item{log_y}{log of the outcome variable \code{y}}
}}
\usage{
Figure15_9
}
\description{
This dataset contains simulated data for Figures 15.9 - 15.11 of Chapter 15.
}
\keyword{datasets}
