% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figures.R
\docType{data}
\name{Figure2_4}
\alias{Figure2_4}
\title{Figure 2.4. Annual Number of Deaths in New York City: Tobacco vs. Other}
\format{A data frame with 591,200 rows and 1 variable:
\describe{
  \item{causes}{cause of death}
}}
\usage{
Figure2_4
}
\description{
This dataset contains data on causes of death in New York City that were used for Figure 2.4 of Chapter 2.
}
\keyword{datasets}
