% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figures.R
\docType{data}
\name{Figure3_2}
\alias{Figure3_2}
\title{Figure 3.2 Data}
\format{A data frame with 100 rows and 1 variable:
\describe{
  \item{fluency}{score on test of Spanish fluency}
}}
\usage{
Figure3_2
}
\description{
This dataset contains simulated test scores of Spanish fluency used to generate Figure 3.2 of Chapter 3.
}
\keyword{datasets}
