% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figures.R
\docType{data}
\name{Figure3_3}
\alias{Figure3_3}
\title{Figure 3.3 Data}
\format{A data frame with 45 rows and 1 variable:
\describe{
  \item{score}{numeric score from rectangular distribution}
}}
\usage{
Figure3_3
}
\description{
This dataset contains simulated scores used to generate Figure 3.3 of Chapter 3.
}
\keyword{datasets}
