% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figures.R
\docType{data}
\name{Figure3_5a}
\alias{Figure3_5a}
\title{Figure 3.5(A) Data}
\format{A data frame with 121 rows and 1 variable:
\describe{
  \item{DistnA}{numeric score from a symmetric distribution}
}}
\usage{
Figure3_5a
}
\description{
This dataset contains simulated scores used to generate Figure 3.5(A) of Chapter 3.
}
\keyword{datasets}
