% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figures.R
\docType{data}
\name{Figure3_5b}
\alias{Figure3_5b}
\title{Figure 3.5(B) Data}
\format{A data frame with 75 rows and 1 variable:
\describe{
  \item{DistnB}{numeric score from a symmetric distribution}
}}
\usage{
Figure3_5b
}
\description{
This dataset contains simulated scores used to generate Figure 3.5(B) of Chapter 3.
}
\keyword{datasets}
