% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figures.R
\docType{data}
\name{Figure3_6and7}
\alias{Figure3_6and7}
\title{Figures 3.6 and 3.7 Data}
\format{A data frame with 69 rows and 2 variables:
\describe{
  \item{NegSkew}{numeric score from a distribution with severe negative skew}
  \item{PosSkew}{numeric score from a distribution with severe positive skew}
}}
\usage{
Figure3_6and7
}
\description{
This dataset contains simulated scores used to generate Figures 3.6 ad 3.7 of Chapter 3.
}
\keyword{datasets}
