% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-figures.R
\docType{data}
\name{Figure5_5}
\alias{Figure5_5}
\title{Figure 5.5 Data}
\format{A data frame with 10 rows and 18 variables:
\describe{
  \item{ax}{days elapsed in a given year}
  \item{ay}{days remaining in that same year}
  \item{bx}{age of elementary school student}
  \item{by}{number of seconds to run a 100-yard dash}
  \item{cx}{introversion score of adolescent boy}
  \item{cy}{aggression score of adolescent boy}
  \item{dx}{moodiness score of college freshman}
  \item{dy}{English ability score of college freshman}
  \item{ex}{weight of male college student}
  \item{ey}{achievement score in statistics of male college student}
  \item{fx}{expected grade in course of college student}
  \item{fy}{course evaluation score given by college student}
  \item{gx}{IQ score of child in grades K – 3}
  \item{gy}{reading achievement score of child in grades K – 3}
  \item{hx}{arithmetic reasoning score of elementary school student}
  \item{hy}{arithmetic fundamentals score of elementary school student}
  \item{ix}{diameter of tree}
  \item{iy}{circumference of tree}
  }}
\usage{
Figure5_5
}
\description{
This dataset contains simulated scores used to generate Figures 5.5(A) - 5.5(I) of Chapter 5.
}
\keyword{datasets}
