% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Hamburger.R
\docType{data}
\name{Hamburger}
\alias{Hamburger}
\title{McDonald's Hamburger Nutrition Information}
\format{A data frame with 5 rows and 4 variables:
\describe{
  \item{name}{type of burger}
  \item{fat}{grams of fat}
  \item{calories}{total calories}
  \item{cheese}{cheese added}
}}
\usage{
Hamburger
}
\description{
This dataset contains the fat grams and calories associated with the different types of hamburger sold by McDonald’s. The data are from McDonald’s Nutrition Information Center.
}
\keyword{datasets}
