% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-IceCream.R
\docType{data}
\name{IceCream}
\alias{IceCream}
\title{Ice Cream Sales Data}
\format{A data frame with 30 rows and 4 variables:
\describe{
  \item{id}{case number}
  \item{temp}{temperature in degrees Fahrenheit}
  \item{barsold}{number of ice cream bars sold}
  \item{relhumid}{relative humidity}
}}
\usage{
IceCream
}
\description{
This dataset contains fabricated data for the temperature, relative humidity, and ice cream sales for 30 days randomly selected between May 15th and September 6th.
}
\keyword{datasets}
