% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leverage.R
\name{leverage}
\alias{leverage}
\title{Leverage}
\usage{
leverage(x)
}
\arguments{
\item{x}{linear regression model given as an \code{\link[stats]{lm}} object.}
}
\value{
A numeric vector of leverage values.
}
\description{
Returns the leverage values for a linear regression model.
}
\examples{
mod = lm(Framingham$SYSBP1 ~ Framingham$TOTCHOL1 + Framingham$AGE1)
leverage(mod)
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[stats]{rstudent}}, \code{\link[stats]{cooks.distance}}
}
