% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Likert.R
\docType{data}
\name{Likert}
\alias{Likert}
\title{Likert-Scale Assertiveness Measure}
\format{A data frame with 30 rows and 1 variable:
\describe{
  \item{Assertiveness}{five-point Likert-scale score of assertiveness, with high scores associated with low levels of assertiveness}
}}
\usage{
Likert
}
\description{
This dataset contains fabricated data for a single survey item measured on a Likert scale. It is given that a survey was administered to 30 individuals and included an item measuring assertiveness by having the individual indicate agreement with the statement: “I have the ability to stand up for my own rights without denying the rights of others.” The response options were: 1 = "strongly agree"; 2 = "agree"; 3 = "neutral"; 4 = "disagree"; 5 = "strongly disagree." Notice that on this scale, high scores are associated with low levels of assertiveness.
}
\keyword{datasets}
