% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Mandext.R
\docType{data}
\name{ManDext}
\alias{ManDext}
\title{Manual Dexterity}
\format{A data frame with 30 rows and 3 variables:
\describe{
  \item{ManualDex}{manual dexterity score}
  \item{Sex}{individual’s sex}
  \item{Treatment}{treatment group assignment}
}}
\usage{
ManDext
}
\description{
This fictional dataset contains the treatment group number and the manual dexterity scores for 30 individuals selected by the director of a drug rehabilitation center. There are three treatments, and the individuals are randomly assigned ten to a treatment. After five weeks of treatment, a manual dexterity test is administered for which a higher score indicates greater manual dexterity.
}
\keyword{datasets}
