% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ManDext2.R
\docType{data}
\name{ManDext2}
\alias{ManDext2}
\title{Manual Dexterity (Dataset #2)}
\format{A data frame with 30 rows and 9 variables:
\describe{
  \item{ManualDex}{manual dexterity score}
  \item{Sex}{individual’s sex}
  \item{Treatment}{treatment group assignment}
  \item{yhat}{predicted outcome for disordinal interaction scenario}
  \item{yhat2}{predicted outcome for ordinal interaction scenario}
  \item{yhat3}{predicted outcome for first no-interaction scenario}
  \item{yhat4}{predicted outcome for second no-interaction scenario}
  \item{yhat5}{predicted outcome for third no-interaction scenario}
  \item{yhat6}{predicted outcome for fourth no-interaction scenario}
}}
\usage{
ManDext2
}
\description{
This is a second fictional dataset that expands on \code{\link[sur]{ManDext}}, adding predicted outcome variables from regression analyses under alternative scenarios.
}
\keyword{datasets}
