% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Marijuana.R
\docType{data}
\name{Marijuana}
\alias{Marijuana}
\title{Marijuana Use of Twelfth Graders}
\format{A data frame with 23 rows and 2 variables:
\describe{
  \item{Year}{year for which data was collected}
  \item{MarijuanaUse}{percentage of twelfth graders who reported that they have ever used marijuana}
}}
\usage{
Marijuana
}
\description{
The dataset contains the year and percentage of twelfth graders who have ever used marijuana for several recent years. The source for these data is The World Almanac and Book of Facts 2014.
}
\keyword{datasets}
