% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.table.R
\name{percent.table}
\alias{percent.table}
\title{Percentage Table}
\usage{
percent.table(x, y = NULL)
}
\arguments{
\item{x}{object containing data for a single variable.}

\item{y}{optional second object to create a contingency table given in percentages. Default setting ignores second object by setting \code{y = NULL}.}
}
\value{
A table of frequency percentages (for one variable) or a contingency table of percentages (for two variables).
}
\description{
For one variable, returns a frequency distribution table given in percentages. For two variables, returns a contingency table given in percentages.
}
\examples{
# frequency table for one variable
percent.table(NELS$region)

# cross-tabulation for two variables
percent.table(Wages$south,Wages$occup)
}
\seealso{
\code{\link[sur]{cumulative.table}}, \code{\link[base]{table}}
}
