% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Politics.R
\docType{data}
\name{Politics}
\alias{Politics}
\title{Gender and Political Party Affiliation}
\format{A data frame with 200 rows and 2 variables:
\describe{
  \item{Gender}{individual's gender}
  \item{Party}{individual's political party affiliation}
}}
\usage{
Politics
}
\description{
This dataset contains data on a fabricated random sample of 200 individuals, 100 females and 100 males, drawn from a population of interest. There are only two variables, both of which are categorical: gender and political party affiliation.
}
\keyword{datasets}
