% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Statisticians.R
\docType{data}
\name{Statisticians}
\alias{Statisticians}
\title{Significant Statisticians}
\format{A data frame with 12 rows and 5 variables:
\describe{
  \item{Statistician}{name of statistician}
  \item{Gender}{gender of statistician, where 1 = “Female” and 2 = “Male”}
  \item{Birth}{year of birth}
  \item{Death}{year of death}
  \item{AmStat}{number of references in \emph{The American Statistician}, 1995-2005}
}}
\usage{
Statisticians
}
\description{
This dataset includes data on 12 statisticians who each have contributed significantly to the field of modern statistics.
}
\keyword{datasets}
