% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Temp.R
\docType{data}
\name{Temp}
\alias{Temp}
\title{Average Monthy Temperatures for Two Cities}
\format{A data frame with 24 rows and 2 variables:
\describe{
  \item{City}{city where temperature was measured}
  \item{Temperature}{average monthly temperature, in degrees Fahrenheit}
}}
\usage{
Temp
}
\description{
This dataset gives the average monthly temperatures (in degrees Fahrenheit) for Springfield, MO and San Francisco, CA.  These data are from Burrill and Hopensperger (1993).
}
\references{
"Exploring Statistics with the T1-81" Burrill, G., & Hopensperger, P. (1993, ISBN:9780201524321)
}
\keyword{datasets}
