% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the.mode.R
\name{the.mode}
\alias{the.mode}
\title{Mode}
\usage{
the.mode(x)
}
\arguments{
\item{x}{object containing data for a single variable.}
}
\value{
A numeric vector of the value(s) of the distribution that have the highest frequency of occurrence.
}
\description{
Function to obtain the mode(s) of a distribution.
}
\examples{
# single mode for factor variable
the.mode(NELS$urban)
# bimodal numeric variable
a = c(14,24,62,12,12,12,36,17,11,99,99,99)
the.mode(a)
}
\seealso{
\code{\link[base]{mean}}, \code{\link[stats]{median}}
}
