% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-UpperBodyStrength.R
\docType{data}
\name{UpperBodyStrength}
\alias{UpperBodyStrength}
\title{Upper Body Strength}
\format{A data frame with 8 rows and 3 variables:
\describe{
  \item{gym}{number of hours spent at the gym weekly}
  \item{strength}{upper body strength score}
  \item{gender}{individual's gender}
}}
\usage{
UpperBodyStrength
}
\description{
This simulated dataset consists of the number of hours eight individuals spend at the gym on a weekly basis along with measures of their upper body strength.
}
\keyword{datasets}
